% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABRCWW.R
\name{ABRCWW}
\alias{ABRCWW}
\title{ABR CWW}
\usage{
ABRCWW(date, CWW_DW = CWW_DW)
}
\arguments{
\item{URL}{the URL string used to query GoogleMaps}
}
\description{
This extracts the ABR business location table from SQL and clips it to CWWs service area
}
\examples{
ABRCWW()

}
\keyword{API}
\keyword{API,}
\keyword{GoogleMapS,}
\keyword{GoogleMaps}
\keyword{Place}

