% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWWClip.R
\name{CWWClip}
\alias{CWWClip}
\title{CWWClip}
\usage{
CWWClip(df = ABRdf)
}
\arguments{
\item{df}{a dataframe from the getABR function}
}
\description{
This clips an ABR dataframe to CWWs service area
}
\examples{
CWWClip()

}
\keyword{ABR,}
\keyword{CWW,}
\keyword{clip}
\keyword{spatial}

