% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getABR.R
\name{getABR}
\alias{getABR}
\title{getABR}
\usage{
getABR(date)
}
\arguments{
\item{URL}{the URL string used to query GoogleMaps}
}
\description{
This extracts the ABR business location table from SQL
}
\examples{
getABR()

}
\keyword{API}
\keyword{API,}
\keyword{GoogleMapS,}
\keyword{GoogleMaps}
\keyword{Place}

