% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGaddGeom.R
\name{PGaddGeom}
\alias{PGaddGeom}
\title{PostgreSQL add geom column}
\usage{
PGaddGeom(db = "ABRdb", tblname, epsg, x, y, geom = "POINT", con = conn)
}
\arguments{
\item{tblname}{name of the table in the database, defaults to ABRdb}

\item{epsg}{what SRID code is the geometry column using? E.g. (EPSG:28355, EPSG:4326 etc.)}

\item{x}{name of latitude column}

\item{y}{name of longitude column}

\item{geom}{the type of the geometry in the datatable E.g. MultiPloygon, MultiPoint, defaults to POINT}

\item{con}{the PostgreSQL connection string}
}
\description{
This adds a geometry column to a data table
}
\examples{
PGaddGeom()

}
\keyword{PostgreSQL,}
\keyword{geom}

