% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGaddSRID.R
\name{PGaddSRID}
\alias{PGaddSRID}
\title{PostgreSQL add SRID}
\usage{
PGaddSRID(tblname, geom, geomcol, epsg, con = conn)
}
\arguments{
\item{tblname}{name of the table in the database}

\item{geom}{the type of the geometry in the datatable E.g. MultiPloygon, MultiPoint}

\item{geomcol}{is the name of the column containing the geometry}

\item{epsg}{what SRID code is the geometry column using? E.g. (EPSG:28355, EPSG:4326 etc.)}

\item{con}{the PostgreSQL connection string}
}
\description{
This adds a spatial reference system to a table
}
\examples{
PGaddSRID()

}
\keyword{PostgreSQL,}
\keyword{SRID,}
\keyword{epsg}

