% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGconstraint.R
\name{PGconstraint}
\alias{PGconstraint}
\title{PostgreSQL add primary key constraint}
\usage{
PGconstraint(tblname, id, con. = conn, drop = FALSE)
}
\arguments{
\item{tblname}{name of the table in the database}

\item{id}{name of the column to set as the primary key for the table}

\item{con.}{is the PostgreSQLConnection, defaults to conn in global environment}

\item{drop}{do you want to remove an existing contraint and drop the column?}
}
\description{
This adds a constrant to a table in PostgreSQL
}
\examples{
PGConstraint()

}
\keyword{PostgreSQL,}
\keyword{constraint}

