% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGremoveduplicates.R
\name{PGdropduplicates}
\alias{PGdropduplicates}
\title{PostgreSQL remove duplicate rows}
\usage{
PGdropduplicates(tblname, id, con. = conn)
}
\arguments{
\item{tblname}{name of the table in the database}

\item{id}{name of the column to set as the primary key for the table}

\item{con.}{is the PostgreSQLConnection, defaults to conn in global environment}

\item{drop}{do you want to remove an existing contraint and drop the column?}
}
\description{
This removes duplicate rows in a PostgreSQL table
}
\examples{
PGdropduplicates()

}
\keyword{PostgreSQL,}
\keyword{constraint}

