% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostgreSQLconnect.R
\name{PostgreSQLconnect}
\alias{PostgreSQLconnect}
\title{PostgreSQL connect}
\usage{
PostgreSQLconnect(dbname = "ABRdb", type = 1, user = NULL, host = NULL)
}
\arguments{
\item{dbname}{name of the database to connect to, default to ABRdb
type connection type, 1 = local database not requiring username/password, 2 = local database requiring username/password, 3 = remote database requiring username/password
user username for user account for database, defaults to NULL
host host for for database, defaults to NULL}
}
\description{
This connects to PostgreSQL server
}
\examples{
PostgreSQLconnect()


}
\keyword{ABRdb}
\keyword{PostgreSQL,}

