% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostgreSQLcreatetable.R
\name{PostgreSQLcreatetable}
\alias{PostgreSQLcreatetable}
\title{PostgreSQL create table}
\usage{
PostgreSQLcreatetable(path, dbname = "ABRdb", type = 1, user = NULL,
  host = NULL)
}
\arguments{
\item{path}{path to the location of the SQL create table scripts
dbname name of the database to connect to, default to ABRdb
type connection type, 1 = local database not requiring username/password, 2 = local database requiring username/password, 3 = remote database requiring username/password
user username for user account for database, defaults to NULL
host host for for database, defaults to NULL}
}
\description{
This creates the ABR tables in the ABRdb database
}
\examples{
PostgreSQLcreatetable()

}
\keyword{ABRdb}
\keyword{PostgreSQL,}

