% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostgreSQLupload.R
\name{PostgreSQLupload}
\alias{PostgreSQLupload}
\title{PostgreSQL upload data}
\usage{
PostgreSQLupload(pathtodata = NULL, pathtoSQL = NULL, tblname = NULL,
  pathtobat = NULL, type, con = con)
}
\arguments{
\item{type}{connection type, 1 = local database not requiring username/password, 2 = local database requiring username/password, 3 = remote database requiring username/password}

\item{path}{path to the location of the SQL create table scripts}

\item{dbname}{name of the database to connect to, default to ABRdb}

\item{user}{username for user account for database, defaults to NULL}

\item{host}{host for for database, defaults to NULL}
}
\description{
This uploads data from the Global Environment and inputs it into a table in PostgreSQL
}
\examples{
PostgreSQLupload()

}
\keyword{ABRdb}
\keyword{PostgreSQL,}

