% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLconnect.R
\name{SQLconnect}
\alias{SQLconnect}
\title{SQLconnect}
\usage{
SQLconnect()
}
\description{
This connects to WVDb1DEVSQL CWW_DW
}
\examples{
SQLconnect()


}
\keyword{CWW_DW}
\keyword{SQL,}
\keyword{WVDb1DEVSQL,}

